Cleaning up cell, adding config options, and .landscape.yml
