+++ templates/config-template.py
@@ -6,3 +6,6 @@ DepthGPIO = '{{ depth_gpio }}'  # GPIO for ultrasonic sensor
 DepthUART = '{{ uart }}'  # UART for ultrasonic sensor
 SerialDev = '{{ serial_dev }}'  # TTY for ultrasonic sensor
 PostURL = '{{ url }}'  # URL for server API endpoint
+RESTART_TIME = '{{ restart_time }}'  # Downtime between cycles
+LOG_PATH = '{{ log_path }}'  # path to log file
+CELL_TIMEOUT = {{ cell_timeout }}  # how many seconds to wait to timeout cell
